chore: Moved test_util.rs to util/test.rs
This commit moves the top-level test_util.rs file under the `util/`
module, and updates callers to use the new path.
