+++ hipcheck/src/util/test.rs
